Refactor: extract the two Context#Key implementations and javadoc
